requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2024.1
cryptography==42.0.7
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.9
django-appconf==1.0.6
django-crispy-forms==2.1
django-cryptography==1.1
django-import-export==3.3.8
django-rest-swagger==2.2.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
html5lib==1.1
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupPy==1.14
MarkupSafe==2.1.5
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.2
oscrypto==1.3.0
Pillow==9.0.1
pycparser==2.22
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tablib==3.5.0
tinycss2==1.3.0
typing_extensions==4.11.0
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.2.1
webencodings==0.5.1
xhtml2pdf==0.2.15
xlrd==2.0.1
xlwt==1.3.0
